rest api for online shop

+++ controllers/feed.js
@@ -4,34 +4,33 @@ const path = require('path');
 const { validationResult } = require('express-validator/check');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
   const perPage = 2;
-  let totalItems; 
-  Post.find().countDocuments()
-  .then(count =>{
-    totalItems = count;
-    return Post.find()
-			.skip((currentPage - 1) * perPage)
-			.limit(perPage);
-  })
-  .then(posts => {
-    res
-			.status(200)
-			.json({
-				message: "Fetched posts successfully.",
-				posts: posts,
-				totalItems: totalItems,
-			});
-  })
-  .catch(err => {
-    if (!err.statusCode) {
-      err.statusCode = 500;
-    }
-    next(err);
-  });
-  
+  let totalItems;
+  Post.find()
+    .countDocuments()
+    .then(count => {
+      totalItems = count;
+      return Post.find()
+        .skip((currentPage - 1) * perPage)
+        .limit(perPage);
+    })
+    .then(posts => {
+      res.status(200).json({
+        message: 'Fetched posts successfully.',
+        posts: posts,
+        totalItems: totalItems
+      });
+    })
+    .catch(err => {
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
+    });
 };
 
 exports.createPost = (req, res, next) => {
@@ -49,18 +48,29 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: 'Maximilian' }
+    creator: req.userId
   });
   post
     .save()
     .then(result => {
+      console.log(result)
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: 'Post created successfully!',
-        post: result
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
     .catch(err => {
@@ -116,6 +126,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -135,29 +150,41 @@ exports.updatePost = (req, res, next) => {
     });
 };
 
-exports.deletePost = ( req,res,next) =>{
+exports.deletePost = (req, res, next) => {
   const postId = req.params.postId;
   Post.findById(postId)
-  .then(post =>{
-    if (!post) {
-      const error = new Error('Could not find post.');
-      error.statusCode = 404;
-      throw error;
-    }
-    clearImage(post.imageUrl);
-    return Post.findByIdAndRemove(postId)
-  })
-  .then(result =>{    
-    console.log(result);
-    return res.status(200).json({messsage:'Deleted Post.'});
-  })
-  .catch(err => {
+    .then(post => {
+      if (!post) {
+        const error = new Error('Could not find post.');
+        error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
+      // Check logged in user
+      clearImage(post.imageUrl);
+      return Post.findByIdAndRemove(postId);
+    })
+    .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
+      res.status(200).json({ message: 'Deleted post.' });
+    })
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
       next(err);
-    })
-}
+    });
+};
 
 const clearImage = filePath => {
   filePath = path.join(__dirname, '..', filePath);
